Refactor multiple choice question components
