+++ client/src/components/MultipleChoiceQuestion/MultipleChoiceAnswer/MultipleChoiceAnswer.svelte
