+++ client/src/components/index.js
@@ -5,7 +5,6 @@ export {Header} from './Header';
 export {Mathjax} from './Mathjax';
 export {LinkedQuestion} from './LinkedQuestion';
 export {Loading} from './Loading';
-export {MultipleChoiceAnswer} from './MultipleChoiceAnswer';
 export {MultipleChoiceQuestion} from './MultipleChoiceQuestion';
 export {Navbar} from './Navbar';
 export {Pseudocode} from './Pseudocode';
